[feat] Implement model-agnostic one-engine eagle3
